@@ -107,13 +107,7 @@ namespace SixLabors.Fonts.IO
         public override bool CanWrite => false;
 
         /// <inheritdoc/>
-        public override long Length
-        {
-            get
-            {
-                throw new NotSupportedException();
-            }
-        }
+        public override long Length => throw new NotSupportedException();
 
         /// <inheritdoc/>
         public override long Position
@@ -145,7 +139,7 @@ namespace SixLabors.Fonts.IO
         {
             // We dont't check CRC on reading
             int read = this.deflateStream.Read(buffer, offset, count);
-            if (read < 1 && this.crcread == null)
+            if (read < 1 && this.crcread is null)
             {
                 // The deflater has ended. We try to read the next 4 bytes from raw stream (crc)
                 this.crcread = new byte[4];
@@ -211,7 +205,7 @@ namespace SixLabors.Fonts.IO
                     this.deflateStream.Dispose();
                     this.deflateStream = null;
 
-                    if (this.crcread == null)
+                    if (this.crcread is null)
                     {
                         // Consume the trailing 4 bytes
                         this.crcread = new byte[4];
