@@ -5,7 +5,7 @@ namespace SixLabors.Fonts.Tables.General.Glyphs
 {
     internal abstract class GlyphLoader
     {
-        public abstract Glyphs.GlyphVector CreateGlyph(GlyphTable table);
+        public abstract GlyphVector CreateGlyph(GlyphTable table);
 
         public static GlyphLoader Load(BinaryReader reader)
         {
