@@ -5,7 +5,7 @@ using System.Numerics;
 
 namespace SixLabors.Fonts.Tables.General.Glyphs
 {
-    internal struct GlyphVector
+    internal readonly struct GlyphVector
     {
         internal GlyphVector(Vector2[] controlPoints, bool[] onCurves, ushort[] endPoints, Bounds bounds)
         {
@@ -17,12 +17,12 @@ namespace SixLabors.Fonts.Tables.General.Glyphs
 
         public int PointCount => this.ControlPoints.Length;
 
-        public Vector2[] ControlPoints { get; private set; }
+        public Vector2[] ControlPoints { get; }
 
-        public ushort[] EndPoints { get; private set; }
+        public ushort[] EndPoints { get; }
 
-        public bool[] OnCurves { get; private set; }
+        public bool[] OnCurves { get; }
 
-        public Bounds Bounds { get; private set; }
+        public Bounds Bounds { get; }
     }
 }
