@@ -1,8 +1,8 @@
 ﻿// Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
+using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Numerics;
 using SixLabors.Primitives;
 
@@ -37,7 +37,7 @@ namespace SixLabors.Fonts
         /// <param name="renderer">The target renderer.</param>
         /// <param name="text">The text.</param>
         /// <param name="options">The style.</param>
-        public static void RenderTextTo(IGlyphRenderer renderer, string text, RendererOptions options)
+        public static void RenderTextTo(IGlyphRenderer renderer, ReadOnlySpan<char> text, RendererOptions options)
         {
             new TextRenderer(renderer).RenderText(text, options);
         }
@@ -49,6 +49,16 @@ namespace SixLabors.Fonts
         /// <param name="options">The style.</param>
         public void RenderText(string text, RendererOptions options)
         {
+            this.RenderText(text.AsSpan(), options);
+        }
+
+        /// <summary>
+        /// Renders the text.
+        /// </summary>
+        /// <param name="text">The text.</param>
+        /// <param name="options">The style.</param>
+        public void RenderText(ReadOnlySpan<char> text, RendererOptions options)
+        {
             IReadOnlyList<GlyphLayout> glyphsToRender = this.layoutEngine.GenerateLayout(text, options);
 
             Vector2 dpi = new Vector2(options.DpiX, options.DpiY);
@@ -57,8 +67,13 @@ namespace SixLabors.Fonts
 
             this.renderer.BeginText(rect);
 
-            foreach (GlyphLayout g in glyphsToRender.Where(x => !x.IsWhiteSpace))
+            foreach (GlyphLayout g in glyphsToRender)
             {
+                if (g.IsWhiteSpace)
+                {
+                    continue;
+                }
+
                 g.Glyph.RenderTo(this.renderer, g.Location, options.DpiX, options.DpiY, g.LineHeight);
             }
 
