@@ -4,9 +4,11 @@ using SixLabors.Fonts.Tests.Fakes;
 
 namespace SixLabors.Fonts.Tests
 {
-    public class FakeFont {
+    public class FakeFont
+    {
         [Fact]
-        public void TestFontMetricProperties() {
+        public void TestFontMetricProperties()
+        {
             Font fakeFont = CreateFont("A");
             Assert.Equal(30, fakeFont.EmSize);
             Assert.Equal(35, fakeFont.Ascender);
@@ -15,8 +17,9 @@ namespace SixLabors.Fonts.Tests
             Assert.Equal(35 - 8 + 12, fakeFont.LineHeight);
         }
 
-        public static Font CreateFont(string text) {
-            FontCollection fc = new FontCollection();
+        public static Font CreateFont(string text)
+        {
+            var fc = new FontCollection();
             Font d = fc.Install(FakeFontInstance.CreateFontWithVaryingVerticalFontMetrics(text)).CreateFont(12);
             return new Font(d, 1);
         }
