@@ -13,18 +13,18 @@ namespace SixLabors.Fonts.Tests.Tables.General
         [Fact]
         public void LoadFormat0()
         {
-            BinaryWriter writer = new BinaryWriter();
+            var writer = new BinaryWriter();
 
-            writer.WriteCMapTable(new []{
-                new SixLabors.Fonts.Tables.General.CMap.Format0SubTable(0, PlatformIDs.Windows, 9, new byte[] { 0, 1, 2 })
+            writer.WriteCMapTable(new[] {
+                new Format0SubTable(0, PlatformIDs.Windows, 9, new byte[] { 0, 1, 2 })
             });
 
             CMapTable table = CMapTable.Load(writer.GetReader());
 
-            Assert.Single(table.Tables.Where(x=>x != null));
+            Assert.Single(table.Tables.Where(x => x != null));
 
             Format0SubTable[] format0Tables = table.Tables.OfType<Format0SubTable>().ToArray();
             Assert.Single(format0Tables);
-        }       
+        }
     }
 }
