@@ -1,9 +1,9 @@
-﻿using System.Numerics;
-using Xunit;
-using SixLabors.Fonts.Tests.Fakes;
-using System.Collections.Immutable;
-using SixLabors.Primitives;
+﻿using System;
 using System.Collections.Generic;
+using System.Numerics;
+using SixLabors.Fonts.Tests.Fakes;
+using SixLabors.Primitives;
+using Xunit;
 
 namespace SixLabors.Fonts.Tests
 {
@@ -72,7 +72,7 @@ namespace SixLabors.Fonts.Tests
             Font font = CreateFont(text);
 
             int scaleFactor = 72 * font.EmSize; // 72 * emSize means 1 point = 1px 
-            RendererOptions span = new RendererOptions(font, scaleFactor)
+            var span = new RendererOptions(font, scaleFactor)
             {
                 HorizontalAlignment = horizental,
                 VerticalAlignment = vertical
@@ -91,6 +91,21 @@ namespace SixLabors.Fonts.Tests
         }
 
 
+        [Fact]
+        public unsafe void MeasureTextWithSpan()
+        {
+            var font = CreateFont("hello");
+
+            Span<char> text = stackalloc char[] { 'h', 'e', 'l', 'l', 'o' };
+
+            int scaleFactor = 72 * font.EmSize; // 72 * emSize means 1 point = 1px 
+
+            SizeF size = TextMeasurer.MeasureBounds(text, new RendererOptions(font, 72 * font.EmSize)).Size;
+
+            Assert.Equal(10, size.Height, 4);
+            Assert.Equal(130, size.Width, 4);
+        }
+
         [Theory]
         [InlineData("h", 10, 10)]
         [InlineData("he", 10, 40)]
@@ -104,10 +119,7 @@ namespace SixLabors.Fonts.Tests
             Font font = CreateFont(text);
 
             int scaleFactor = 72 * font.EmSize; // 72 * emSize means 1 point = 1px 
-            SizeF size = TextMeasurer.MeasureBounds(text, new RendererOptions(font, 72 * font.EmSize)
-            {
-
-            }).Size;
+            SizeF size = TextMeasurer.MeasureBounds(text, new RendererOptions(font, 72 * font.EmSize)).Size;
 
             Assert.Equal(height, size.Height, 4);
             Assert.Equal(width, size.Width, 4);
@@ -117,7 +129,7 @@ namespace SixLabors.Fonts.Tests
         public void TryMeasureCharacterBounds()
         {
             string text = "a b\nc";
-            GlyphMetric[] expectedGlyphMetrics = new GlyphMetric[] {
+            var expectedGlyphMetrics = new GlyphMetric[] {
                 new GlyphMetric('a', new RectangleF(10, 0, 10, 10), false),
                 new GlyphMetric(' ', new RectangleF(40, 0, 30, 10), false),
                 new GlyphMetric('b', new RectangleF(70, 0, 10, 10), false),
@@ -127,10 +139,10 @@ namespace SixLabors.Fonts.Tests
             Font font = CreateFont(text);
 
             int scaleFactor = 72 * font.EmSize; // 72 * emSize means 1 point = 1px 
-            Assert.True(TextMeasurer.TryMeasureCharacterBounds(text, new RendererOptions(font, 72 * font.EmSize), out IReadOnlyList<GlyphMetric> glyphMetrics));
+            Assert.True(TextMeasurer.TryMeasureCharacterBounds(text, new RendererOptions(font, 72 * font.EmSize), out GlyphMetric[] glyphMetrics));
 
-            Assert.Equal(text.Length, glyphMetrics.Count);
-            for (int i = 0; i < glyphMetrics.Count; i++)
+            Assert.Equal(text.Length, glyphMetrics.Length);
+            for (int i = 0; i < glyphMetrics.Length; i++)
             {
                 GlyphMetric expected = expectedGlyphMetrics[i];
                 GlyphMetric actual = glyphMetrics[i];
@@ -188,8 +200,8 @@ namespace SixLabors.Fonts.Tests
             Font font = c.Install(TestFonts.SimpleFontFileData()).CreateFont(12);
 
             int scaleFactor = 72 * font.EmSize; // 72 * emSize means 1 point = 1px 
-            GlyphRenderer glyphRenderer = new GlyphRenderer();
-            TextRenderer renderer = new TextRenderer(glyphRenderer);
+            var glyphRenderer = new GlyphRenderer();
+            var renderer = new TextRenderer(glyphRenderer);
             renderer.RenderText(text, new RendererOptions(new Font(font, 1), 72 * font.EmSize, new PointF(x, y)));
 
             Assert.Equal(expectedX, glyphRenderer.GlyphRects[0].Location.X, 2);
@@ -198,7 +210,7 @@ namespace SixLabors.Fonts.Tests
 
         public static Font CreateFont(string text)
         {
-            FontCollection fc = new FontCollection();
+            var fc = new FontCollection();
             Font d = fc.Install(new FakeFontInstance(text)).CreateFont(12);
             return new Font(d, 1);
         }
